recipes-kernel: Copy to msm_default only when it doesn't exist
When devtool utility is used to build, the copy to msm_default
fails. This is because, copy is already done when source code
was checked out to workspace. Hence, copy to msm_default only
when it doesn't exist.

+++ recipes-kernel/linux/linux-qcom-custom_6.6.bb
@@ -145,7 +145,8 @@ python copy_msm() {
     src = os.path.join(d.getVar('S'), "drivers/gpu/drm/msm")
     dst = os.path.join(d.getVar('S'), "drivers/gpu/drm/msm_default")
 
-    shutil.copytree(src, dst)
+    if not os.path.exists(dst):
+        shutil.copytree(src, dst)
 }
 do_unpack[postfuncs] += "copy_msm"
